pyproject.toml
[project]
name = "agentic-rag-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2026.2.25",
    "fastapi==0.104.1",
    "httpx==0.28.1",
    "langgraph==0.0.20",
    "motor==3.3.2",
    "openai>=1.50.0",
    "pydantic>=2.12.5",
    "pymongo==4.6.0",
    "pymupdf==1.27.1",
    "pypdf2==3.0.1",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.6",
    "qdrant-client==1.7.0",
    "uvicorn[standard]==0.24.0",
]
